resolves for comments
