+++ envVariables.js
@@ -129,10 +129,6 @@ let enviromentVariables = {
     "message" : "App Portal base url",
     "optional" : true,
     "default": "https://dev.sunbirded.org"
-  },
-  "SUNBIRD_SERVICE_URL" : {
-    "message" : "Required sunbird service url",
-    "optional" : false
   }
 }
