+++ generics/constants/endpoints.js
@@ -31,7 +31,7 @@ module.exports = {
     LIST_PROJECT : "/v1/userProjects/list",
     GET_QUESTIONS : "/v1/solutions/questions",
     GET_OBSERVATION : "/v1/observations/details",
-    GET_LOCATION_DATA : "/api/data/v1/location/search",
-    GET_FORM_DATA : "/api/data/v1/form/read",
-    GET_SCHOOL_DATA : "/api/org/v1/search"
+    GET_LOCATION_DATA : "/v1/location/search",
+    GET_FORM_DATA : "/v1/form/read",
+    GET_SCHOOL_DATA : "/v1/org/search"
 }
